requirements.txt
--index-url=https://pypi.tuna.tsinghua.edu.cn/simple

django==2.2
mysqlclient==2.1.0
pycryptodome==3.9.7
